// src/generated/models/Cr2b6_shiftassignmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr2b6_shiftassignmentscr2b6_shift = {
  566210000: 'A',
  566210001: 'B',
  566210002: 'C',
  566210003: 'D'
} as const;
export type Cr2b6_shiftassignmentscr2b6_shift = keyof typeof Cr2b6_shiftassignmentscr2b6_shift;
export const Cr2b6_shiftassignmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr2b6_shiftassignmentsstatecode = keyof typeof Cr2b6_shiftassignmentsstatecode;
export const Cr2b6_shiftassignmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr2b6_shiftassignmentsstatuscode = keyof typeof Cr2b6_shiftassignmentsstatuscode;

export interface Cr2b6_shiftassignmentsBase {
  "cr2b6_Employee@odata.bind"?: string;
  cr2b6_enddate?: string;
  cr2b6_id: string;
  cr2b6_note?: string;
  cr2b6_shift?: Cr2b6_shiftassignmentscr2b6_shift;
  cr2b6_shiftassignmentid: string;
  cr2b6_startdate?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr2b6_shiftassignmentsstatecode;
  statuscode?: Cr2b6_shiftassignmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr2b6_shiftassignments extends Cr2b6_shiftassignmentsBase {
  cr2b6_employeename?: string;
  cr2b6_shiftname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr2b6_employee?: object;
  _cr2b6_employee_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}