// src/generated/models/Cr2b6_peoplesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr2b6_peoplesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr2b6_peoplesstatecode = keyof typeof Cr2b6_peoplesstatecode;
export const Cr2b6_peoplesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr2b6_peoplesstatuscode = keyof typeof Cr2b6_peoplesstatuscode;

export interface Cr2b6_peoplesBase {
  cr2b6_fullname: string;
  cr2b6_peopleid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr2b6_peoplesstatecode;
  statuscode?: Cr2b6_peoplesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr2b6_peoples extends Cr2b6_peoplesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}